UnionFind now does not need the root in find(). Adjusted kruskal accordingly

+++ template.cc
@@ -17,7 +17,7 @@ typedef vector<vii> vvii;
 #define mp make_pair
 #define mt make_tuple
 #define ge(a, b) get<(b)>((a))
-#define SZ(x) (x).size()
+#define SZ(x) (int)(x).size()
 #define ft front()
 #define bk back()
 #define endl '\n'
